try fix scopes
